requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.25
django-cors-headers==4.1.0
django-filter==23.5
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
future==1.0.0
idna==3.7
importlib-metadata==6.7.0
kombu==5.2.4
Markdown==3.4.4
MarkupPy==1.14
numpy==1.21.6
odfpy==1.4.1
openpyxl==3.1.2
pandas==1.3.5
Pillow==9.5.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.1
pyzk==0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
typing-extensions==4.7.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
